increase metadata to int size for GT compat

+++ src/main/java/zmaster587/libVulpes/block/BlockMeta.java
@@ -31,7 +31,7 @@ public class BlockMeta {
 		return block;
 	}
 
-	public byte getMeta() {
+	public int getMeta() {
 		if(meta != WILDCARD)
 			return (byte) meta;
 		return 0;
